backend_expenses/requirements.txt
celery
redis
black>=24.0,<25.0
isort>=5.13,<6.0
flake8>=7.0,<8.0
pre-commit>=3.7,<4.0
drf-yasg>=1.21.7

pytest-xdist>=3.6,<4.0
pydantic>=2.0,<3.0
django-filter>=23.0,<24.0
ipython>=8.0,<9.0
django-extensions
DJango>=4.2,<5.0
djangorestframework>=3.15,<4.0
gunicorn>=22.0,<23.0
psycopg2-binary>=2.9,<3.0
python-decouple>=3.8,<4.0
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
factory_boy>=3.3,<4.0
djangorestframework-simplejwt>=5.3,<6.0
